+++ ports/mynewt/include/memfault_shell.h
@@ -1,0 +1,16 @@
+#pragma once
+
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int mflt_shell_init(void);
+
+#ifdef __cplusplus
+}
+#endif